A bunch of changes in order to facilitate:
i) Specification of command-line options for plugins via
-plugin:pluginName[par1=x,par2=y] syntax (requires each plugin to
implement a setParam(par1,x) function, which handles the possible
command-line inputs).

ii) Modification of output filenames for the Postprocess
plugins based on the command line parameters given to the 
ModelExtension plugins. For example, specifying epsilon=1.0 now 
results in the StructTrace output being written to a file
with ending ".eps_"+fixedEpsilonValue+".struct";