Indel moves are now handled via the coreModel object, which is of McmcModule type. This makes it very easy to add, remove and combine different moves for the indel parameters, specified in the initCoreModel() function.
Among the new moves specified here are some reparameterised versions of
the (lambda, mu) moves, to try to deal with the high correlation between
lambda and mu. 

Also, the proposal scheme for these parameters has been
changed from a
uniform random walk to a logistic or multiplicative scheme. Uniform
could also be added on to the list of possible options.

The plan is to now move the other sampleX() functions for the core model
into the coreModel, which will allow, for example, easy addition or
removal of new topology moves in the future.