Included an option to include an initial randomisation period, during which all coreModel moves are accepted provided that they do not result in impossible parameter values. During this period, the topologyWeight is currently bumped up to 100, so we're mainly randomising the initial tree.
The length of the randomisation period (default=0) can be set to 500
with syntax like that shown below

java -jar statalign.jar -mcmc=1m,1m,100,500 seqs.fasta

The fourth parameter (the 500) is optional. If in parallel mode then it
expects the parallel parameters to be the last argument, regardless of
whether the randomisation period is set.

The randomisation period can also be set via the GUI in the MCMC
settings dialogue box.

The acceptances in the randomisation period are not used for automatic
tuning of proposal variances.