Modifications such that imputation now only inserts and deletes columns when actually necessary.
However, there appears to be a problem, because the order of some
columns seems to be changing before and after the move, most easily seen
from the full printed alignments. Currently investigating why this may
be (most probably a call to prev/next is done in the wrong order or
something along those lines).