Changed command line option for switching on debug mode -- now requires -debug=true. This is not listed in the usage information, because presumably nobody reading that would need to switch on debug mode...
AlignmentMove efficiency improved by not recalculating quantities that are already cached.

Alignment proposals implemented that use variable mu and lambda, equal to the current values multiplied by a constant which is a member variable of the AlignmentMove object.

Restoration after five-way topology move now keeps cached information for lower part of the tree.