Changed command line option for switching on debug mode -- now requires -debug=true. This is not listed in the usage information, because presumably nobody reading that would need to switch on debug mode...
    AlignmentMove efficiency improved by not recalculating quantities
that are already cached.
    
    Alignment proposals implemented that use variable mu and lambda,
equal to the current values multiplied by a constant which is a member
variable of the AlignmentMove object.

	Also enabled autoTuning for alignment moves in second half of
burnin, which seems to result in improvements in accepted change rates
(i.e.
acceptance rates for moves that change the alignment) with much smaller
window sizes. One of the moves is kept at a fixed expected window
length, to make sure that more global realignments are also possible.

	More variable alignment proposals are only switched on in second half
of burnin, because if they are introduced before the alignment has
properly converged then convergence will take a lot longer. They are
intended mainly for mixing after convergence.
    
    Restoration after five-way topology move now keeps cached
information for lower part of the tree.