Postprocessing information is now output on-the-fly, and is flushed along with current MCMC info every Mcmc.LOG_INTERVAL samples (default 100).
This enables the chain to be terminated early in command-line mode
whilst having recorded everything up to the previous multiple of
LOG_INTERVAL samples. 

In GUI mode, all the output files will be handled correctly. In
command-line mode, the tree file will not have the correct NEXUS
terminus added to the end, so the following will need to be added some
other way in order to create a valid NEXUS file:

END; [Trees]
BEGIN st_Assumptions;
	treestransform=TreeSelector;
	splitstransform=EqualAngle;
	SplitsPostProcess filter=dimension value=4;
	autolayoutnodelabels;
END; [st_Assumptions]