Implemented track for B-factors and marginal probabilities of current alignment.
Added computation of S+ marginal probabilities, and made these the
default annotation for the current alignment (currently not outputted
with the MPD alignment).

StructAlign now continues with warning when localEpsilon is selected but
no B-factors provided.

Changed start of tree file to be 'NEXUS' rather than 'nexus', for
compatibility with other software.