Fix to spike mixture prior (relevant only to runs carried out using the 'sigmaSpike' option, such that 0<gamma<1).
Previously a contribution from the prior was skipped in the Hastings
ratio, resulting in a bias in the estimation of the mixture proportion,
gamma. This has been fixed, and the default Beta prior modified to
achieve a similar amount of shrinkage.

Various modifications have also been made to the sampling of nu during
the burn-in, to prevent getting stuck in local modes when sampling with
the spike mixture prior.