Plugin panels now no longer show in GUI when screenable=false.
In StructAlign mode, tree with edge lengths weighted by structural
diffusivity is printed to file (if the option is selected), and shown in
the GUI.

Removed 'proposed' columns from StructAlign parameter output, since this
information was rarely useful, and increased the size and complexity of
the output.

Some renaming of StructAlign output files.