Overwriting of sequences now only occurs for data types that are not RawSequences, so that the order of PDB and Fasta file input does not affect the behaviour.
In theory, if two non-RawSequence data types were to be read in
separately, each with a different sequence for the same taxon name, then
the order of command line input would still affect which of the
sequences gets assigned to the taxon. However, there are currently no
cases where the user would want to do that intentionally.