Preliminary testing of MPI version.
Temperature exchanges seem to proceed ok. Currently most of the plugins
are not outputting correctly, hence not possible to assess the output at
this point in time. 

Currently the classpath is altered manually from within the program, to
work around the fact that it's being initiated via MPJ's starter.jar.
This also means that inter-plugin dependencies are not correctly
resolved at the moment.

No reason to fix this yet, since we may eventually not want to invoke
via starter.jar.